Rename  services/SQLite_example.py -> services/sqlite_database_handler.py
Fix handlers/tour_selection_form_handlers.py
